Add user_tutors table for storing tutors.
Instead of using the user_course table to store tutors, I've created a
new table user_tutors to store tutors. There are built-in assumptions in
the rest of the code that user_course ONLY stores instructors, it would
probably break stuff if we keep using user_course to store both
instructors and tutors.

The element 'courses/list_instructors' was moved to
'list/unordered_list_users' since it's now also used to list tutors.
Since the element takes in an array of users to display already, it was
already generalized enough to not be instructor specific anymore.