changed the filter for tutors to use a function
Previously, if statements were used to filter the group members. However
this can become pretty troublesome becuase we can easily miss a "spot".
We now have a function to grab all group members and another to grab only
students in the group. This way we can have two separate variables which
we can use whenever we need to.