fixed the import users
The form previously did not have the correct indices for the courses.
That has now been resolved by using asort instead of sort.

The list of ids of users being created/updated is now being created by
looping through the results instead of building it after each save. The
reason is if the user already exist, we cannot register them in another
course using import (before), because the user is "not created" therefore
would not appear in the "insertedIds" array.

The last major problem was that if there are two or more users in the csv
files that already exist, their information woudl be overwritten with the
information from the last user in the csv file. The reason is the variable
$tmp is used twice in the function for different purposes. I have changed
one of the variables to $temp instead.