fixed import group
The problem was the UserEnrol and GroupsMember models' functions cannot
be accessed for some reasons. We needed those functions to help us check
validity of the groups in the csv file.

Nathaniel and I have changed the code to use the User and Group models
instead. We have also added a check to see if Tutors are in the course
by checking the Tutor array as well.

Some other problems were just invalid variables (variables that don't
exist anymore) and some spelling erros.

In this commit, I have also fixed a minor bug I introduced in my last
commit where new users are not being enrolled in the classes. This was
due to an invalid variable.