fixed import users - no roles assigned
Before when we import users, no roles are assigned to them (roles_users
table). To fix this I had to assign a 'Role' array to the results array.
By doing this, I also had to put the user's information in an User array
in the result, therefore I had to make some adjustments due to new indices
being introduced.

All users imported are defaulted to have a student role, because the page
is linked to "Import Students" on the course home page.