Remove the role column from the user table.
Since we should be using the RolesUsers table instead for storing roles.

This was prompted by trying to refactor the fixtures to take the table
layout from the default database instead of having to manually specify
the tables. The enum() values for the roles column turns out tobe
unsupported by SimpleTest since enum() is a MySQL specific feature.

A lot of old code is still using the old User.roles column. Most of
which doesn't seem to be used anymore, so I removed them. The ones that
were still being used was modified to work with the new RolesUsers
table.