edited the test cases for sys_function
function getTopAccessibleFunction has been removed from the model
and the test cases because the function is no longer being used.
It was probably used when the role field in the user table was used
to keep track of the users' roles.

function getAllAccessibleFunction has also been removed from the
model and the test cases because the only function that calls it
from sys_container controller is no longer in use.

Lastly, the test case (the only one left) is using imported data from
the sample data.