fixed bug when a course has no groups
When a course has no groups, a bug would appear where an array, that
is created for checking what is the next available group num, would
not exist. This causes many problems for php functions that require
a defined array, that's not empty.

I've fixed it by checking if any groups exist. If not, the next
defined group num is 1.