edited the users api and wrote the groups api
For users, the api has been edited to check for http requests using
isDelete for example instead of checking with the facts of whether
data is given and/or and id is given. Currently, the GET requests
prints out the user(s)' id, username, role id, first name and last
name. We may want to edit if we want to provide more data or less
data. For POST, I've only asked for username, first name, and last
name. We can again come back to edit this (eg. password).

I've also changed the POST, PUT, and DELETE views. If POST or
PUT requests are successful, the id of the user is printed in json.
If DELETE is successful, nothing is printed on the view. We are
going to implement error messages if the requests are unsuccessful.

For groups, I've also used the same structure as user where we are
using an "if else" structure checking for which http request is
requested. For POST, groups, group name, group num, course id, and
group members are asked. While GET, only returns id, group number,
group name, and course id are returned.

For both users and groups, we should probably implement some
validations such as data for all fields are provided. We should also
checked for whether group numbers, course ids, group ids, and etc
are valid before processing the requests.