made a slight change to the api
I've made slight change to the api for courses (POST/PUT) to not
only display the id (number only) but also the "tag" id so that
the user can know what the number means.

I've also cleaned up the interface by removing the broken images
of the rounded corner bottoms (left/right).