removed unused views and edited some permission checking
I've removed index.ctp and search.ctp for evaluations because they
are no longer being used.

In the evaluation controller, I have started to edit the permission
checking mechanism. Before the function "noStudentsAllow" was used
to prevent student users to access those sections of the
application. I have changed some of them to hasPermission.

For evaluations/export, I have also checked whether the user is an
instructor or admin of the course. For some other functions such
as sys_parameters and sys_functions, only superadmins can access
them, because currently it is the only role that has the permission
for those controllers.