changed all permission checking to hasPermission for evaluations
All the functions that require authentication now uses hasPermission.

I've also fixed the bug for viewGroupSubmissionDetails (invalid indices).
I think the problem was the function that was used to generate the
list of group members was edited and has removed the fields that the
submission details page need. Therefore I have just modified the array
in the controller.