Refactored events/add.
Moved from a table based layout to a regular CSS styled form layout.
Reduced the JavaScript complexity by handing the form validation off to
the Event model instead. Also simplified save by changing the fields so
that they take advantage of the CakePHP's automagic for saving related
models.

The Penalty system was simplified to it's base form, which is just how
many % to deduct from the total mark each day that it is late. With the
last day being the mark that will be deduced from then onwards.

Dropped the prototype based date picker in favour of the jquery one. Had
to import jquery-ui and jquery-ui-timepicker as a result.