edited functions/tests/sample data to use new penalty structure.
We have "simplified" the penalty structure. There is now only one
type instead of two. Therefore the function/test for penaltyType
has been removed.

The change is the finalPenalty is now the "first day" it is used
instead of a negative number (eg. -1, -2). For example, if the
final deduction is in effect starting on the 4th day, the days
late for the final deduction will be 4 instead of -1 or -2.