fixed the radio button's css for emailtemplates/add
Before, the radio buttons & label were not aligned with the rest
of the fields of the form. We have fixed it by clearing the left
and right of the fieldset - which caused the indentation, that we
don't want.

I have also removed the commented out/old code from the penalty
model.