edited the evaluations/export to display all events in the course
Previously, the export function has been changed to only take an
event id, which broke the link from the course home page.

To solve this problem, I have changed the function to take two
parameters instead. The new parameter is the word 'course' or
'event'. The url is now evaluations/export/course/$id for
example.

If 'course' is chosen, the page allows the user to choose any
events that are in the course ($id). If 'event' is chosen,
the event would be selected and can't be changed.