fixed survey results view for instructor
Previously, there were two main problems. One was for checkboxes, only
one of the choices is checked off rather than all the choices checked
off by the student. The second problem was when a student don't respond,
the code will give error of invalid indices.

To fix the two problems, I redid the whole view. Instead of using
html form tags to create the results, I've instead just echoed
the questions and responses. For multiple choice and checkbox
questions, all choices are printed. The choices that were chosen
is in bold green. For the single line and multi line answers,
they are just echoed (italicized). If the answer is too long
(especially for the multi-line answers), the answer will automatically
break into multiple lines (or a scrollbar will appear).