fixed the student view for survey results
I've fixed the broken view of survey results for students. It now
looks almost identical to the instructor view of individual survey
results.

I've also had to fixed the student_index.ctp file to have the link
to the results again. Now Surveys will appear on student index
"forever" because it doesn't have a result release beginning and
end dates.