fixed some minor bugs
1) from the list of actions for courses, I've removed View Instructor,
because it isn't working and it wouldn't make sense for it to work
if the course has two or more instructors. The user can easily view
the information by clicking on the instructors' names from the home
page

2) fixed the faculties/view. There was an error where the index
"User" cannot be found. I found out the problem was the array
provided does not have the array User or any of the fields
being displayed. I fixed it by adding an extra line to look
up the admins' information from the user's table.

3) fixed the link surveygroups/viewresult/ from the course home
page. The page was displaying fine, but none of the dates from
the submissions were printing. The problem was that no $eventId
was specified when it isn't specified in the url.

4) When creating groups manually, the group_num is one less than
it suppose to be. I've fixed that by starting the count from 1
instead of 0 since the first group number is 1.

5) The simple evaluation preview's total points available to be
distributed is 0 (for 3 users). I've fixed that by not setting
the value to 0. Instead I initialized it to be 3 (3 test users)
times the "point_per_member" value.

6). I've fixed the css for courses and users add/edit which broke
when we implemented some css for events add/edit. I've fixed it by
putting the Events class name in front of each label type for
events. This revert courses and users add/edit back to normal.