Add superadmin to sql files.
The superadmin used to be created by the install_controller. This worked
fine when we didn't need to refer to the superadmin in the sql template
files, or at least, when such references weren't strictly enforced.

The faculty admin permissions feature has relegated the superadmin to be
a faculty admin that's in all faculties. This means we need to add the
superadmin to all faculties. The user_faculties table, however, enforces
relations, which means that when we created the database, it rejected
the superadmin as a non-existent user. The solution was to add an entry
for superadmin into the sql templates.

The install_controller, instead of creating the superadmin, now justs
modifies the existing entry for the superadmin. Note that this patch
also changes the superadmin creation to properly use the User model
instead of raw sql queries.