implemented the faculty admin permissions for many of the controllers
I've added a few functions to some of the models to help generate
list of courses, events, etc that are required to generate the list
the admins/ super admins can view, add, edit, and delete.

Faculty admins have access to courses belonging to departments in
their faculty/faculties and the elements within those courses.
Super admin is basically a "faculty admin" that has access to
all faculties.

I've added a function (getMyDepartmentsCourseList) to the user model.
It is used to generate all the course the user (admin/super admin)
has access to based on which faculties (and departments) they have
access to. There is one parameter, which is defaulted to list.

It calls the findAllyByUserId in the UserFaculty model to find
all the faculties the user has access to. With this list, we pass
it as a parameter to the getByUserFaculties in the department model
to generate a list of departments the user has access to. This is
then passed to getByDepartments in the course model (new function,
explained below), which generates the list of courses.

As I have explained above, getByDepartments() returns the list of
coures the user has access to. It takes two parameters. One is
the list of departments and the other is $findType, so that
we can have the flexibility of a list or "all".

Lastly, I have tried to add the word "Error:" before each failure
message, so that it will sound less personal/offensive.