fixed the evaluations export - no instructors
The export evaluation results interface has the option to
display the instructors' names in the csv/excel file. However,
no names are displayed whether the option is selected or not.

The problem was it was never implemented. I have now implemented
to print all the instructors in the header if the option is selected.