made some minor fixes/adjustments
1) In both the basic ipeer.sql and sample data sql files, I've
added a field "availability" to the simple evaluations table.
This is to allow users to have the option to make their template
public or private. I would be adding that option to the simple eval
add/edit form later.

2) I've added the permission for admins to have access to the
evaluations controller, which for some reasons was not given
before.

3) I've fixed some css bugs for the user and events form. There
seemed to be a small problem where a change for events form
affected the users form as well. This is now fixed.