finished implementing the faculty admin access for the other controllers
In general, the rule I follow is if the element has a course id
associated with it, I will give the user permission if they
have access to the course. One example would be surveys.
If not course ids are associated, such as the evaluations,
the permissions are based on the creator's id. If the creator
teaches a course that the current user has access to, then
the current user can access the element.

Along the way, I also had to make minor adjustments to the sql
files, to remove the functions/email permission from admin
because the permissions are now based on which faculties they
have access to.

I've also added an availability column for simple evaluation
in the "All My Tools" page to reflect the option I've added
in a previous commit.