fixed checkstyle problems - mostly missing documents
Most of the problems was missing documentation for functions in
the controller and the variables not being public in the model.
These problems have now been fixed.

I have also fixed the problem in the user model where an extra
parameter was documented for a function. It has now been removed.