Add OAuth nonce storage for the API.
Nonces are supposed to be unique. The only way we can enforce that is if
we have a way of storing nonces we've already seen. This table stores
nonces we've seen. Since it's infeasible to store every single nonce
we've seen, there's an expiration time for each nonce. We combine this
with timestamps to guarantee a unique nonce within a certain time period
while rejecting all requests that are outside of this time period.

Also had to move the insert user entries up, as mysql was rejecting our
oauth entries since the root user didn't exist.

Fixed some minor inconsistencies between the two sql files.