Require OAuth verification for API calls.
Added OAuth checking to beforeFilter() in v1_controller, so every single
request now must go through OAuth verification before executing. Note
that role based restrictions have not been applied yet.

There is an empty oauth() method in v1_controller that is used just for
checking that OAuth requests are being accepted, it is the url used by
the v1 test case to check that we're following OAuth standards.

OAuth verification has been split into 3 parts:

Check Required Parameters:
The OAuth protocol requires certain parameters to be present. This
section also checks that we don't have requests trying to use
unsupported signature methods.

Check Signature:
Check that the calculated HMAC-SHA1 signature of this message matches
the signature given. Note that there is a security issue here to be
fixed later - the signature doesn't include the content of the http
request being transmitted during POST and PUT requests. So those
requests are still vulnerable to MITM attacks that the signature is
supposed to defeat.

Check Nonce:
Check that the nonce is unique and that the timestamp is within
allowable tolerances of our own. We only guarantee the uniqueness of
nonces within the allowed tolerance, so timestamps that are too far off
from ours will cause the request to be rejected.

Note that running the v1_controller tests now require the OAuth PECL
extension to be installed. The OAuth PECL extension is used to check
that our implementation of OAuth is correct.