fixed super admin access & instructors access to users/index
- instructors are now unable to go to the users index. They can
acces their students profiles by going to the class list instead

- super admin access to various elements of ipeer such as
evaluation templates have been fixed to be able to view
everything. The difference is super admin can view private
evaluation templates created by faculty admin for example. This
change was also made so that the ajaxList would not need to
process potentially hundreds of course ids for the filters.

- I have created two new permissions user/index and superadmin.

- user/index is to allow users to access users/index

- superadmin is to allow superadmins to access everything