fixed some minor bugs
- fixed the error when a user first log in. It was already fixed
in the "create_acl" file. However the change was not placed
in the sql files. This has now been done

- changed the expiry date of the sample token to year 2032
because 32 bit computers will give the year 2092 a timestamp
of 0.

- changed the redirect in courses/add for access denied to
courses/index because it would be more convenient for instructors
if they stay on the index page. Students and tutors would be
redirected again to home anyways

- removed self-enrolment information from the courses view

- commented out the Search tab from navigation because the functions
are not used anymore

- changed the values for availability for email templates to 0
and 1 instead of private and public because 0 and 1 are used
in the database

- changed the availability field for email templates view from 0 or
1 to radio buttons (public or private). The reason is it would be
easier to understand.

- added an availability field to the simple evaluation view
page

- in editProfile, the date format has been changed to YYYY-MM-DD for
the tokens' expiry dates

- adding user email notification checkbox's position wasn't right
in internet explorer, so I fixed it in ipeer.css

- made some minor adjustments to the css for oauth tokens in
editProfile, to accomodate for longer keys

- changed the implementation of the returned url for user add and edit
to a more stable implementation