removed pop-ups and fixed events add bug
- All pop-ups are removed from the application because if the user
opens them in IE, they would be logged out of the application.
Instead the "pop-ups" are now new tabs or windows.

- events add had a bug where the wrong variable was set to populate
the list of courses. Also when a super admin try to access the page,
an invalid variable error comes up, which has now been fixed.
This bug was introduced when I fixed the code to allow super
admins to view everything and not depend on faculties and
departments.

- The same bug was introduced into survey add, which has now been
fixed.

- the emailer write view also had a css problem in IE, where
the label of the fields are centred in their cell, while in the other
browswers they are aligned left. This has now been fixed.