fixed some minor bugs and fixed a permission loophole
- removed the LOM column from the mixeval index because it is not
used in the new form (vs version 2)

- fixed user/import. The additional csv allowed type that fixed
the "windows" problem broke the application for mac. The reason
was the additional csv type overwrite the original. To fix this
problem we placed the two csv types in the same array - instead
of two separate arrays

- fixed an undefined field problem in the mixeval model. It was just
a simple spelling error

- for some reason the email merge field broke for the email
template. It has now been fixed by adding the onChange attribute
to the dropdown menu

- removed the space right before the "Comment" for oauth tokens and
clients in the editProfile

- fixed the problem with removing questions from mixeval in IE. The
problem was a console.log was in the code. It has now been removed.

- fixed the permission checking where students can retake
evaluations and surveys if they know the exact url. Now, if
a submission is made and the submitted field in the submissions
table is true, then access would be denied.

- fixed the problem where admins are added as instructors if they
created the course. Now only users that are not admins would
be automatically added.