fixed some minor bugs and implemented some minor features
- removed the extra $joinTableCreator from the emailer's ajax.
It isn't exactly a bug, except it overwrites the correct one
and cause the text to say evaluation instead of email

- reimplemented two filters for surveys that were not entirely
taken off. They are for filtering the user's survey and
surveys for a specific course.

- implemented the instructors field for course add so that admins
can add instructors to the course at the time of course creation

- fixed the bug of the emailer making two copies of an email
even though the single email option was chosen. The reason
was that the value for the option was changed to 2 from 1.
This has now been fixed.