removed permission from emailer and fixed redirect in users view
- removed permission checking for emailing to individual users.
Since only instructors, faculty admins, and super admins can
access the emailer, we don't need to check for permission since
they can email all users in the system.

- fixed the redirect from users view. This does not matter for
admins and super admins because they can view all users. However
it does not make sense for instructors to be redirected to users
index and then to home each time they try to veiw a user they
have no access to. So instead I have changed it to redirect
to whichever page the user was before. This also helps
keeps the error message which is gone when the user is redirected
twice.