added import users to the api
- I've added the functionality where users can import an array of
users. The view will display all users successfully added to the
system or already exists (determined by the combination of username,
first_name, and last_name).

- I've also changed the _oauthReq function in the test file
for the api to display more information if an error occurs. The
reason for the change is so that we can use it to get the
statuscode to verify delete works.

- some minor adjusts have also been made to reflect the input and
output we want for the api.