API enrol student/instructor/tutor in course.
Also get enrolment information in a course.

Note that only two required parameters are needed to enrol a user to a
course. The user's username and the role_id that the user is to assume.
v1_controller uses the role_id to put the user in the correct table,
either UserCourse, UserEnrol, or UserTutor.

The return value, if everything went right, is the same JSON string sent
to iPeer. If an error occured, the HTTP error code will include a
message stating which user caused the error.

Getting the enrolment information returns the user's username and
role_id. The id field is useless and should be removed in a future
commit.

Note that we're using usernames in lieu of user ids because it's easier
to access the username in the iPeer Building Block.