API remove users from course.
Takes the same parameters as adding a user to a course, except this is
an HTTP DELETE instead of PUT.

Since the adding user to a course code was basically copied and pasted
as a basis for the removing users part, the two really should somehow be
merged into one section. I can't think of a nice way to do this though,
that doesn't involve nesting too deeply.

Note that due to continuing problems with UserEnrol's methods being
confused with SQL queries, I've moved a lot of it's methods into the
User model. With corresponding reductions in UserEnrol's code and tests.
And additional tests for the User model.