Refactor grades API, re-enable grades by user.
Getting grades by event id was modified to return username instead of a
user id for user identification. This broke the tests for retriving
grades from a single user by id. This commit is to fix that.

Retriving grades for an individual user is now by username instead of by
user id, to be more in line with retriving grades by event id. The tests
were modified to match.

The grades API implementation was also rewritten to be simpler and less
nested.