API: Fix OAuth tests, implement groupMembers.
GroupMembers was still using the placeholder implementation used for
quick testing instead of the full version. I changed the routes to go to
the full version and fixed it up to be compatible with usernames instead
of user_id. Also fixed the delete implementation and test cases.

Looks like the OAuth error outputs were changed, the test cases
expectations now matches the new outputs.