Revert use of submitter_id in eval simple.
EvaluationSimple model was trying to use 'submitter_id' when it
should've been using 'evaluator'.

While I do remember talking about changing evaluator to submitter_id, I
unfortunately can't remember exactly why we were considering that move.
But since the change wasn't reflected in the database, I'm going to
assume that we actually decided not to do that and that this was a
remnant piece of code that accidentally got it.