Refactor user summary and course submenu view.
elements/courses/submenu was unnecessarily nested with
submenu_items.ctp, so I brought in the code from submenu_items.ctp and
removed it.

views/users/user_summary.ctp was using a div when it really should've
been using a header, so removed the unnecessary div and CSS class and
changed it to a h3.