Refactor evaltools to simplify tables/improve nav
views/evaltools/index.ctp Instead of just a huge giant table of all the
various evaluations and email templates, split the table up into
individual tables for each eval/email type that'll be easier to maintain.

The Survey and Email Templates tables actually weren't being shown due
to it using the old way of checking permissions. It might be a good idea
to permission control the other tables as well, but for now, I assume
that if you have access to the EvalTools page, you have access to
everything else as well.