SimpleEvalResult fix comments, css refactor.
This is for view_simple_evaluation_result which instructors and admins
see.

Comments release wasn't working due to broken parameter checking. The
check was removed and it seems to be working now.

Turns out that the usage of tables on this page is necessary. So the
code wasn't seriously simplified. I've removed the accordion stuff
though, as it's completely unecessary to have all that extra complexity
when scrolling down a longer page is much more intuitive and has better
cross-platform compatibility.

All the tables are using the standardtable css class so that we can cut
down on the number of custom table CSS we've currently got.