CSS: Nicer fonts, change color theme for tables.
Now imports 2 fonts from Google Web Fonts.
- Open Sans, for the headers, since it allows a font-weight of 600
- Lato, for everything else

http://www.google.com/webfonts/specimen/Open+Sans
http://www.google.com/webfonts/specimen/Lato

Open Sans is there purely because it seems to look nicer when bolded. I
originally was going to use the 2 different levels of bold available in
Open Sans to set the headers more apart from each other, but didn't
bother in the end. Open Sans also seem to retain a bit more readability
at smaller font sizes, so I'm using it for the user's name and logout
button in the banner too.

The tables were using a lot of blue, it seems that there might be too
much blue in iPeer, so I changed the colors for standardtable to orange.
But then realized that the orange was a bit too dark and makes stuff a
bit hard to read. So now, I'm using lighter yellow/brown/orange-ish
colours.

Unfortunately, because themeing the jQuery based datatables takes a bit
more effort due to the increased complexity (and some image assets), I
only did a minimal customization. Enough so that it fits in with the new
colours, but not enough so that I have to change the more complicated
assets.