Fix SimplEvalResult can't find full_name in user.
The various user variables used on the page is loaded by
GroupsMembers->getGroupMembers which simply uses a join query to get
user data. Unfortunately, since full_name is a virtual field provided by
the User model and not stored in the database, this means that
view_simple_evaluation_results.ctp doesn't get the 'full_name' field.
This is why it was concatting first_name and last_name all over the
place.

I've replaced the getGroupMembers call with a replacement that actually
uses the User model. Judging from the state of the evaluation component,
I'm guessing that there's other evaluation result pages that might
suffer from this same bug, will have to check those out later.