Group submission detail remove element, rerelease.
I can't figure out what reReleaseEvaluation is supposed to do, so I've
just removed it entirely.

Group submission detail for some reason was unnecessarily taken out into
an element, I've removed the element and moved the code into the regular
view. The code was refactored to use the standardtable style and use
html helper to generate the group members table.

Note that the problem with using the Group models to get User
information as noted before exists here too and was temporarily patched.
A better refactor is needed later.