fixed top buttons css problem in Firefox
For some reason the css style float:right was causing the margin
values to be ignored. Therefore I fixed it by fremoving the float
style but instead aligned the links/buttons to the right, which
creates the same effect and the margins are used in Firefox as
well.

Note: Ticket 398