Installer: fix debug, move ldap to optional, css.
Debug panel fails in the installer because the User model isn't
initialized during installation (no User table yet). Added checks so
that it only attempts to get user information if the User model is
present.

The PHP LDAP module check should be in the Optional Requirements section
since it's only necessary if they want to use LDAP authentication.

Change the CSS for .help so that it only applies to the installer div
and it's more visible.