Student Home: Upcoming/Submitted/Expired events.
Determining what surveys are show to students turns out to be a bit more
complicated than I thought due to all the different dates we've got.

Some definitions to show what I'm trying to achieve:

Release period - is the time interval in which students are allowed to
make submissions to the event. This is stored by the release_date_begin
and release_date_end fields in the Event table. And also the is_release
virtual field in the Event model.

Result release period - is the time interval in which students are
allowed to view resulting data from the event. This is stored by the
result_release_date_begin and result_release_date_end fields in the
Events table. And also by the is_result_release virtual field in the
Event model.

Upcoming events - Events that are still within their release period and
which the user haven't made any submissions to.

Submitted events - Events that the user has made a submission to and
hasn't gone past the result release period.

Expired events - Events that the user has not made a submission to and
hasn't gone past the result release period.

Once an event is past its result release period, it will no longer show
up in the student view.

Note that since surveys doesn't have a result release period, I'm using
the release period instead.

Also added a 'summary statistic' second at the top of the student home
page. It'll show the total number of events that needs to be done if
there are any overdue. Standard red, orange, green color background is
used to indicate if the user has anything to worry about.