Add/View/List surveys now same as eval events.
You can now List/View/Add surveys in the same way as
Simple/Rubric/Mixevals. This was actually easier to do than creating a
whole separate facility to deal with surveys separately. Since the plan
was to ultimately have surveys behave the same way as all other events
anyways.

Editing surveys will be done in the next commit.

Note that the split between survey and survey template still needs to be
done on the template side, as the survey template creation is still
creating an entry in the event table.