Allow editing survey events, fix event template id
Following on from the previous commit, can now edit survey events.
Mostly just a matter of getting the same changes in add into edit.

While implementing this, I discovered a bug where when editing an event,
the correct template of the existing event isn't selected. This was an
oversight due to there being no evaluation types with multiple templates
until surveys came along.

There is now a mechanism to make sure that the template of the existing
event is selected instead of the default first template.